A Function That Writes, Saves, and Exports Syntax for Fitting Latent Actor-Partner Interdependence Models

This function takes the outputted object from dyadVarNames() and automatically writes, returns, and exports (.txt) lavaan() syntax for specifying Actor-Partner Interdependence Models (APIMs). Users can also invoke configural, loading, and/or intercept invariant measurement models, and particular types of structural comparisons.

apimSEM(dvn, lvxname, lvyname, model = "configural", compare = "none")

Arguments

dvn
input object from dyadVarNames()

lvxname
input character to (arbitrarily) name LV X in lavaan syntax

lvyname
input character to (arbitrarily) name LV Y in lavaan syntax

model
input character used to specify which level of invariance is modeled. Defaults to "configural"

compare
input character to specify which type of structural parameters ("actor" = actor effects, "partner" = partner effects, "all_effects" = actor and partner effects, x_means, y_means, or all_means) are constrained to equivalencey between partners. Default is "none".

Value

character object of lavaan script that can be passed immediately to lavaan functions. Users will receive message if structural comparisons are specified when the recommended level of invariance is not also specified. If user supplies dvn with containing X or Y variables, they are alerted to respecify the dvn object.

See also

Examples

dvn = dyadVarNames(dat, xvar="X", yvar="Y", sep = ".",distinguish1 = "1", distinguish2 = "2")
apim.script.config = apimSEM(dvn, lvxname = "Conflict",
lvyname = "Satisfaction", model = "configural")
apim.script.load = apimSEM(dvn, lvxname = "Conflict",
lvyname = "Satisfaction", model = "loading")
apim.script.int = apimSEM(dvn, lvxname = "Conflict",
lvyname = "Satisfaction", model = "intercept")
apim.script.load.actor = apimSEM(dvn, lvxname = "Conflict",
lvyname = "Satisfaction", model = "loading", compare = "actor")