A Function That Computes Effect Sizes of Dyadic Noninvariance

This function takes a model written with dyadCFA() and fitted with lavaan() and returns d_MACS values for each indicator variable. Function is a renamed version of my dMACS_long function (https://github.com/jsakaluk/dMACS_long), which was adapted from Johannes Karl's (https://johanneskarl.netlify.com) between-groups dMACS function (https://github.com/Jo-Karl/dMACS) with his assistance.

dyInvarEs(fit.cfa)

Arguments

fit.cfa
output from a dyadCFA() configural invariance model fitted with lavaan()

Value

numeric vector of dMACS values for each indicator of a given LV

See also

Examples

dvn <- dvn2(DRES, x_order = "sip", x_stem = "sexsat", x_delim2=".", x_item_num="\\d+", distinguish_1="1", distinguish_2="2")

#> Error in dvn2(DRES, x_order = "sip", x_stem = "sexsat", x_delim2 = ".",     x_item_num = "\\d+", distinguish_1 = "1", distinguish_2 = "2"): could not find function "dvn2"

config.script = dyadCFA(dvn, lvname = "SexSat",  model = "configural")

#> Error in dyadCFA(dvn, lvname = "SexSat", model = "configural"): could not find function "dyadCFA"

fit <- lavaan::cfa(config.script, data = DRES, std.lv = FALSE,
auto.fix.first= FALSE, meanstructure = TRUE)

#> Error in lavaan::lavaan(model = config.script, data = DRES, std.lv = FALSE,     auto.fix.first = TRUE, meanstructure = TRUE, model.type = "cfa",     int.ov.free = TRUE, int.lv.free = FALSE, auto.fix.single = TRUE,     auto.var = TRUE, auto.cov.lv.x = TRUE, auto.cov.y = TRUE,     auto.th = TRUE, auto.delta = TRUE, auto.efa = TRUE): object 'config.script' not found

dydMACS(con.fit.config)

#> Error in lavaan::lavInspect(fit.cfa, what = "rsquare"): object 'con.fit.config' not found