A Function That Exports a Table of Dyadic Invariance Model Comparison Indices

This function takes the model from fitted dyCFA() scripts and exports a table of expected output.

dyMIOutput(config, load, int, robust = F)

Arguments

config
fitted dyadic configural invariance model lavaan object

load
fitted dyadic loading invariance model lavaan object

int
fitted dyadic intercept invariance model lavaan object

robust
logical for whether a robust estimator has been used (default is F)

Value

Prints word files for the table outputed to the users working directory

See also

Examples

dvn <- dyadVarNames(dat, xvar="X",  sep = ".",
distinguish1 = "1", distinguish2 = "2")

con.config.script = dyadCFA(dvn, lvname = "Conflict",  model = "configural")

#> Error in dyadCFA(dvn, lvname = "Conflict", model = "configural"): could not find function "dyadCFA"

con.loading.script = dyadCFA(dvn, lvname = "Conflict",  model = "loading")

#> Error in dyadCFA(dvn, lvname = "Conflict", model = "loading"): could not find function "dyadCFA"

con.intercept.script = dyadCFA(dvn, lvname = "Conflict",  model = "intercept")

#> Error in dyadCFA(dvn, lvname = "Conflict", model = "intercept"): could not find function "dyadCFA"

con.fit.config <- lavaan::cfa(con.config.script, data = dat, std.lv = FALSE, auto.fix.first= FALSE, meanstructure = TRUE)

#> Error in lavaan::lavaan(model = con.config.script, data = dat, std.lv = FALSE,     auto.fix.first = TRUE, meanstructure = TRUE, model.type = "cfa",     int.ov.free = TRUE, int.lv.free = FALSE, auto.fix.single = TRUE,     auto.var = TRUE, auto.cov.lv.x = TRUE, auto.cov.y = TRUE,     auto.th = TRUE, auto.delta = TRUE, auto.efa = TRUE): object 'con.config.script' not found

con.fit.loading <- lavaan::cfa(con.loading.script, data = dat, std.lv = FALSE, auto.fix.first= FALSE, meanstructure = TRUE)

#> Error in lavaan::lavaan(model = con.loading.script, data = dat, std.lv = FALSE,     auto.fix.first = TRUE, meanstructure = TRUE, model.type = "cfa",     int.ov.free = TRUE, int.lv.free = FALSE, auto.fix.single = TRUE,     auto.var = TRUE, auto.cov.lv.x = TRUE, auto.cov.y = TRUE,     auto.th = TRUE, auto.delta = TRUE, auto.efa = TRUE): object 'con.loading.script' not found

con.fit.intercept <- lavaan::cfa(con.intercept.script, data = dat, std.lv = FALSE, auto.fix.first= FALSE, meanstructure = TRUE)

#> Error in lavaan::lavaan(model = con.intercept.script, data = dat, std.lv = FALSE,     auto.fix.first = TRUE, meanstructure = TRUE, model.type = "cfa",     int.ov.free = TRUE, int.lv.free = FALSE, auto.fix.single = TRUE,     auto.var = TRUE, auto.cov.lv.x = TRUE, auto.cov.y = TRUE,     auto.th = TRUE, auto.delta = TRUE, auto.efa = TRUE): object 'con.intercept.script' not found

dyMIOutput(con.fit.config, con.fit.loading, con.fit.intercept, robust = FALSE)

#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'fitMeasures': object 'con.fit.config' not found