Helpers for creating output files

Helpers for creating output files

makeTable(dvn, fit, model, tabletype)

makeFigure(fit, type)

Arguments

dvn
input object from dyadVarNames()

type
input character to indicate whether parameters "free" or "equated" in estimation

param
input character for what kind of parameter is being scripted ("act", "apim_part", "mim_part", "cf")

lvar
input character for whether scripting helpers target latent "X" or :Y" indicator variables in dvn

lvxname
input character to (arbitrarily) name LV X in lavaan syntax

lvyname
input character to (arbitrarily) name LV Y in lavaan syntax

partner
input character to indicate parameters for first or second dyad member

See also