A Function That Exports Tables and/or SEM Digrams based on dySEM models

This function takes the model from fitted dySEM() scripts and exports table(s) and/or a path diagram figure of expected output.

outputModel(
  dvn,
  model = NULL,
  fit,
  table = TRUE,
  tabletype = NULL,
  figure = TRUE,
  figtype = NULL,
  writeout = FALSE
)

Arguments

dvn

input dvn list from scrapeVarCross

model

character input specying type of model to output (e.g., "cfa", "apim", "cfm")

fit

input object from fitted lavaan model

table

logical input of whether table output is desired. Default is TRUE

tabletype

character input of what type of table(s) is(are) desired. options are "measurement" (i.e,, loadings, intercepts,), "structural" (i.e., latent slopes, such as actor/partner effects, k parameters), or "both" (i.e., both measurement and structural tables)

figure

logical input of whether figure output is desired. Default is TRUE

figtype

character input of what type of figure is desired

writeout

logical (default is FALSE) for whether outputted table and/or figure should be written to an output subdirectory in current working directory

Value

Ignore console (prints unneccessary semPlot::semPaths details). More importantly, prints word files for the table(s) and/or figure, outputed to the users working directory

Examples

dvnx <- scrapeVarCross(dat = commitmentQ, x_order = "spi", x_stem = "sat.g", x_delim1 = ".",
x_delim2="_", distinguish_1="1", distinguish_2="2")
sat.config.script <-  scriptCFA(dvnx, lvname = "Sat", model = "configural")
sat.config.mod <- lavaan::cfa(sat.config.script, data = commitmentQ, std.lv = FALSE,
auto.fix.first= FALSE, meanstructure = TRUE)
outputModel(dvnx, model = "cfa", fit = sat.config.mod, table = TRUE,
tabletype = "measurement", figure = "TRUE", figtype = "standardized")
dvnxy <- scrapeVarCross(dat = commitmentQ, x_order = "spi", x_stem = "sat.g",
x_delim1 = ".", x_delim2="_", distinguish_1="1", distinguish_2="2",
y_order="spi", y_stem="com", y_delim1 = ".", y_delim2="_")
apim.script.config <-  scriptAPIM(dvnxy, lvxname = "Sat", lvyname = "Com", model = "configural")
apim.config.fit <- lavaan::cfa(apim.script.config, data = commitmentQ, std.lv = FALSE,
auto.fix.first= FALSE, meanstructure = TRUE)
outputModel(dvnxy, model = "apim", fit = apim.config.fit, table = TRUE,
tabletype = "measurement", figure = "TRUE", figtype = "standardized")