Helper-functions for scraping variable names and data set properties

Helper-functions for scraping variable names and data set properties

sipExtractor(dat, stem, delim1, item_num, delim2, distinguish)

spiExtractor(dat, stem, delim1, item_num, delim2, distinguish)

spitExtractor(dat, stem, delim1, item_num, delim2, distinguish, delim3, wave)

siptExtractor(dat, stem, delim1, item_num, delim2, distinguish, delim3, wave)

Arguments

dat
input data frame of indicators of a particular LV

stem
input character stem of indicator variables for LV

delim1
optional input character of first delimiting character

item_num
optional input character to indicate range of variable numbers (defaults to scrape any digits)

delim2
optional input character of second delimiting character

distinguish
input character used as the identifier a particular partner

delim3
optional input character of second delimiting character

wave
input character for the particular wave of a longitudinal design

See also