A Function That Writes, Saves, and Exports Syntax for Fitting the I-NULL model for indistinguishable dyads

This function takes the outputted object from scrapeVarCross() and automatically writes, returns, and exports (.txt) lavaan() syntax for the I-NULL model described in Olsen & Kenny (2006)

Usage

scriptINULL(dvn, lvxname = "X", lvyname = NULL, writescript = FALSE)

Arguments

dvn

input dvn list from scrapeVarCross

lvxname

input character to (arbitrarily) name X LV in lavaan syntax

lvyname

(optional) input character to (arbitrarily) name Y LV in lavaan syntax

writescript

input logical (default FALSE) for whether lavaan script should be concatenated and written to current working directory (in subdirectory "scripts")

Value

character object of lavaan script that can be passed immediately to lavaan functions

See also

Examples

dvn <- scrapeVarCross(dat = DRES, x_order = "sip", x_stem = "PRQC", x_delim1 = "_",
x_delim2=".", x_item_num="\\d+", distinguish_1="1", distinguish_2="2")
qual.inull.script <- scriptINULL(dvn, lvxname = "Qual")