Ratings of relational satisfaction and commitment from 282 (M)ixed-sex couples

A data set containing 5 ratings of satisfaction and 5 ratings of commitment for each member of a mixed-sex romantic dyad. Measured using the Investment Model Scale subscales (Rubult, Martz, & Agnew, 1998). Data are from Sakaluk, Fisher, and Kilshaw's (in press) study of dyadic invariance. Variable names have been re-coded to follow a stem-item-partner ("sip") order, with a delimiter ("_") between the item number and partner distinguishing character.

data(commitmentM)

Format

A data frame with 282 rows and 20 variables:

sat.g1_f

Satisfaction item 1 for female partner

sat.g2_f

Satisfaction item 2 for female partner

sat.g3_f

Satisfaction item 3 for female partner

sat.g4_f

Satisfaction item 4 for female partner

sat.g5_f

Satisfaction item 5 for female partner

com1_f

Commitment items item 1 for female partner

com2_f

Commitment items item 2 for female partner

com3_f

Commitment items item 3 for female partner

com4_f

Commitment items item 4 for female partner

com5_f

Commitment items item 5 for female partner

sat.g1_m

Satisfaction item 1 for male partner

sat.g2_m

Satisfaction item 2 for male partner

sat.g3_m

Satisfaction item 3 for male partner

sat.g4_m

Satisfaction item 4 for male partner

sat.g5_m

Satisfaction item 5 for male partner

com1_m

Commitment items item 1 for male partner

com2_m

Commitment items item 2 for male partner

com3_m

Commitment items item 3 for male partner

com4_m

Commitment items item 4 for male partner

com5_m

Commitment items item 5 for male partner

References

Sakaluk, J. K., Fisher, A. N., & Kilshaw, R. E.(in press). Dyadic measurement invariance and its importance for replicability in romantic relationship research. Personal Relationships.