A Function That Writes, Saves, and Exports Syntax for Fitting Latent Mutual influence Model

This function takes the outputted object from dyadVarNames() and automatically writes, returns, and exports (.txt) lavaan() syntax for specifying Mutual Influence Models (MIMs). Users can also invoke configural, loading, and/or intercept invariant measurement models, and particular types of structural comparisons.

scriptMIM(
  dvn,
  scaleset = "FF",
  lvxname,
  lvyname,
  model = "configural",
  equate = "none",
  k = FALSE
)

Arguments

dvn
input object from dyadVarNames()

scaleset
input character to specify how to set the scale of the latent variable(s). Default is "FF" (fixed-factor; see Details for rationale), but user can specify "MV" (Marker Variable)

lvxname
input character to (arbitrarily) name LV X in lavaan syntax

lvyname
input character to (arbitrarily) name LV Y in lavaan syntax

model
input character used to specify which level of invariance is modeled. Defaults to "configural"

equate
input character to specify which type of structural parameters ("actor" = actor effects, "partner" = partner effects, "all_effects" = actor and partner effects, x_means, y_means, or all_means) are constrained to equivalencey between partners. Default is "none". Actor or partner effect constraints require at least a loading-invariant model to be specified, otherwise a warning is returned; mean constraints require an intercept-invariant model to be specified, otherwise a warning is returned.

k
input logical for whether Kenny & Ledermann's (2010) k parameter should be calculated to characterize the dyadic pattern in the mim. Defaults FALSE, and requires at least a loading-invariant model to be specified, otherwise a warning is returned.

Value

character object of lavaan script that can be passed immediately to lavaan functions. Users will receive message if structural comparisons are specified when the recommended level of invariance is not also specified. If user supplies dvn with containing X or Y variables, they are alerted to respecify the dvn object.

See also

Examples

dvn <- scrapeVarCross(dat = DRES, x_order = "sip", x_stem = "PRQC", x_delim1 = "_", x_delim2=".", x_item_num="\\d+", distinguish_1="1", distinguish_2="2",
                    y_order="sip", y_stem="sexsat", y_delim2=".", y_item_num="\\d+")
mim.script.config = scriptMIM(dvn, lvxname = "Quality",
lvyname = "SexSat", model = "configural", scaleset = "MV")
mim.script.load = scriptMIM(dvn, lvxname = "Quality",
lvyname = "SexSat", model = "loading", scaleset = "MV")
mim.script.int = scriptMIM(dvn, lvxname = "Quality",
lvyname = "SexSat", model = "intercept", scaleset = "MV")
mim.script.load.actor = scriptMIM(dvn, lvxname = "Quality",
lvyname = "SexSat", model = "loading", equate = "actor")